Moving proximal operators back to tenalg
